feat: 战斗选定对手时递增目标计数
1. 按原版16位递增、CF保留和三个物理调用位置接入目标计数。
2. 覆盖回绕、flags、typed-stop与父级后缀，并推进战斗清单至307/422。

+++ src/battle/legacy_battle_actor_runtime_reset.cpp
@@ -262,7 +262,7 @@ void synchronize_actor_write(
     copy_overlap_to_owner(
         actor.residual->bytes_0174_029f.data(),
         0x0174U,
-        actor.residual->bytes_0174_029f.size(),
+        static_cast<u32>(actor.residual->bytes_0174_029f.size()),
         image,
         offset,
         size
@@ -270,7 +270,7 @@ void synchronize_actor_write(
     copy_overlap_to_owner(
         actor.residual->bytes_0d34_0d4f.data(),
         0x0D34U,
-        actor.residual->bytes_0d34_0d4f.size(),
+        static_cast<u32>(actor.residual->bytes_0d34_0d4f.size()),
         image,
         offset,
         size
@@ -351,7 +351,9 @@ void synchronize_actor_write(
         copy_overlap_to_owner(
             actor.group_b_configuration->profile_buffer.data(),
             kProfileBase,
-            actor.group_b_configuration->profile_buffer.size(),
+            static_cast<u32>(
+                actor.group_b_configuration->profile_buffer.size()
+            ),
             image,
             offset,
             size
@@ -361,7 +363,7 @@ void synchronize_actor_write(
                 actor.group_b_composition->action_text.data()
             ),
             kActionTextBase,
-            actor.group_b_composition->action_text.size(),
+            static_cast<u32>(actor.group_b_composition->action_text.size()),
             image,
             offset,
             size
